# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HelpDesk Pro API - Multitenant support platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics
      security:
        - cookieAuth: []
      parameters:
        - name: tenantId
          in: query
          schema:
            type: integer
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
  /dashboard/tickets-by-status:
    get:
      operationId: getTicketsByStatus
      tags: [dashboard]
      summary: Get tickets grouped by status
      security:
        - cookieAuth: []
      parameters:
        - name: tenantId
          in: query
          schema:
            type: integer
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Tickets by status
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StatusCount"
  /dashboard/tickets-by-priority:
    get:
      operationId: getTicketsByPriority
      tags: [dashboard]
      summary: Get tickets grouped by priority
      security:
        - cookieAuth: []
      parameters:
        - name: tenantId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Tickets by priority
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PriorityCount"
  /dashboard/tickets-over-time:
    get:
      operationId: getTicketsOverTime
      tags: [dashboard]
      summary: Get ticket volume over time
      security:
        - cookieAuth: []
      parameters:
        - name: tenantId
          in: query
          schema:
            type: integer
        - name: period
          in: query
          schema:
            type: string
            enum: [day, week, month]
            default: month
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Tickets over time series
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimeSeriesPoint"
  /dashboard/tickets-by-technician:
    get:
      operationId: getTicketsByTechnician
      tags: [dashboard]
      summary: Get tickets grouped by assigned technician
      security:
        - cookieAuth: []
      parameters:
        - name: tenantId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Tickets by technician
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TechnicianCount"
  /dashboard/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity feed
      security:
        - cookieAuth: []
      parameters:
        - name: tenantId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
  /dashboard/top-categories:
    get:
      operationId: getTopCategories
      tags: [dashboard]
      summary: Get most frequent ticket categories
      security:
        - cookieAuth: []
      parameters:
        - name: tenantId
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Top categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryCount"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
      required:
        - email
        - password
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        tenantId:
          type: integer
          nullable: true
        tenantName:
          type: string
          nullable: true
        tenantSlug:
          type: string
          nullable: true
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - role
        - active
        - createdAt
...
    DashboardStats:
      type: object
      properties:
        totalTickets:
          type: integer
        newTickets:
          type: integer
        openTickets:
          type: integer
        resolvedTickets:
          type: integer
        closedTickets:
          type: integer
        avgResolutionHours:
          type: number
          nullable: true
        totalTenants:
          type: integer
        totalUsers:
          type: integer
        totalTechnicians:
          type: integer
        pendingTickets:
          type: integer
        urgentTickets:
          type: integer
      required:
        - totalTickets
        - newTickets
        - openTickets
        - resolvedTickets
        - closedTickets
        - totalTenants
        - totalUsers
        - totalTechnicians
        - pendingTickets
        - urgentTickets
    StatusCount:
      type: object
      properties:
        status:
          type: string
        count:
          type: integer
        label:
          type: string
      required:
        - status
        - count
        - label
    PriorityCount:
      type: object
      properties:
        priority:
          type: string
        count:
          type: integer
        label:
          type: string
      required:
        - priority
        - count
        - label
    TimeSeriesPoint:
      type: object
      properties:
        date:
          type: string
        created:
          type: integer
        resolved:
          type: integer
      required:
        - date
        - created
        - resolved
    TechnicianCount:
      type: object
      properties:
        userId:
          type: integer
        userName:
          type: string
        count:
          type: integer
        resolved:
          type: integer
      required:
        - userId
        - userName
        - count
        - resolved
    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        action:
          type: string
        entityType:
          type: string
        entityId:
          type: integer
        entityTitle:
          type: string
          nullable: true
        userId:
          type: integer
        userName:
          type: string
        tenantName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - action
        - entityType
        - entityId
        - userId
        - userName
        - createdAt
    CategoryCount:
      type: object
      properties:
        category:
          type: string
        count:
          type: integer
      required:
        - category
        - count
...